imp: make OAuthToken work moving Gson config
