+++ common/src/main/java/org/openbaton/nfvo/common/configuration/SystemConfiguration.java
@@ -17,20 +17,16 @@
 
 package org.openbaton.nfvo.common.configuration;
 
-import com.google.gson.Gson;
-import com.google.gson.GsonBuilder;
 import org.apache.catalina.Context;
 import org.apache.catalina.connector.Connector;
 import org.apache.tomcat.util.descriptor.web.SecurityCollection;
 import org.apache.tomcat.util.descriptor.web.SecurityConstraint;
-import org.openbaton.catalogue.nfvo.messages.Interfaces.NFVMessage;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.context.embedded.EmbeddedServletContainerFactory;
 import org.springframework.boot.context.embedded.tomcat.TomcatEmbeddedServletContainerFactory;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Scope;
 
 /** Created by lto on 10/11/15. */
 @Configuration
@@ -39,15 +35,6 @@ public class SystemConfiguration {
 
   @Value("${nfvo.https:false}")
   private boolean https;
-
-  @Bean
-  @Scope("prototype")
-  Gson gson() {
-    return new GsonBuilder()
-        .setPrettyPrinting()
-        .registerTypeAdapter(NFVMessage.class, new GsonDeserializerNFVMessage())
-        .create();
-  }
 
   @Bean
   public EmbeddedServletContainerFactory servletContainer() {
